Use and extend existing statsd sink class
