+++ source/server/config/stats/BUILD
@@ -37,7 +37,7 @@ envoy_cc_library(
         "//source/common/config:well_known_names",
         "//source/common/network:address_lib",
         "//source/common/network:resolver_lib",
-        "//source/common/stats:dog_statsd_lib",
+        "//source/common/stats:statsd_lib",
         "//source/server:configuration_lib",
     ],
 )
