+++ source/server/config/stats/statsd.cc
@@ -26,7 +26,7 @@ Stats::SinkPtr StatsdSinkFactory::createStatsSink(const Protobuf::Message& confi
         Network::Address::resolveProtoAddress(statsd_sink.address());
     ENVOY_LOG(debug, "statsd UDP ip address: {}", address->asString());
     return Stats::SinkPtr(
-        new Stats::Statsd::UdpStatsdSink(server.threadLocal(), std::move(address)));
+        new Stats::Statsd::UdpStatsdSink(server.threadLocal(), std::move(address), false));
     break;
   }
   case envoy::api::v2::StatsdSink::kTcpClusterName:
