+++ test/common/stats/BUILD
@@ -52,20 +52,6 @@ envoy_cc_test(
 )
 
 envoy_cc_test(
-    name = "udp_dog_statsd_test",
-    srcs = ["udp_dog_statsd_test.cc"],
-    deps = [
-        "//source/common/network:address_lib",
-        "//source/common/network:utility_lib",
-        "//source/common/stats:dog_statsd_lib",
-        "//test/mocks/stats:stats_mocks",
-        "//test/mocks/thread_local:thread_local_mocks",
-        "//test/test_common:environment_lib",
-        "//test/test_common:network_utility_lib",
-    ],
-)
-
-envoy_cc_test(
     name = "thread_local_store_test",
     srcs = ["thread_local_store_test.cc"],
     deps = [
